refactor(desktop): simplify MessageScrollbackRail into single self-contained file
Delete over-engineered model file and tests. Inline helpers, drop
custom types, and use CSS truncate instead of JS string slicing.
